Move kernel extension helpers to new trait
